Changed chat to poll files instead of DB ongoing
Major change to the back end of chat.
When a user logs in their chat file is cleared.
Chat history is retrieved from the DB
When chats are sent they are inserted in the DB and echoed to each user's chat file.
The long polling to get chats monitors the file.
Note this design can allow for users to direct-message each other or facilitate the implementation of channels in the future.