Changes for major release 12 May 2018
We're finally ready for the next update, our last one (glaring bugs aside) on cross-tables as we'll be focusing on moving to our new server.

The most noticeable change will be cosmetic. The idea behind which is to make the site warmer and a little more inviting to people who aren't used to word study. There are basic colour shifts animations/fades in places and we settled on green/black at the general colour scheme going forward. Just about everything you see has been reworked one way or another.

Behind the scenes, large chunks of code have been rewritten and more bug fixes than we're capable of remembering have gone in (we'll be logging these better in the near future). In any case, here are some highlights.

---
Invaders:
--

This has had a major rewrite behind the scenes. The changes you'll notice are that the scale adjusts to different screen sizes when loaded (This may need more fine tuning, let us know) and there's a much needed method underneath for tracking the alphagrams.

---
Leaderboard:
---

You'll notice some thin strips of different colours towards the top of the leaderboard. These are awards. These have arrived a little prematurely as we're not going to start tracking daily and weekly until Monday.

How do awards work?

Top spot at the end of day,week,month, year always gets 'Emerald'.

If you're in the top 25%, you get bronze (minimum 5 users for it to appear),

Top 12.5% silver (minimum 9 users for it to appear)

Top 5% gold (minimum 21 users for it to appear).

As the user base increases and certain achievements devalue, new ones appear... For example, 'Sapphire' kicks in now at top 1% (at least 101 users) on yearly for Ed Rossiter... and it just goes on like that.

Each timeframe also gives a different number of award points.
Daily:1, Weekly:3, Monthly:5, Yearly:30.

The first wave of planned updates on xerafin.net includes basic user profiles, so it won't be long before you'll be able to see individual totals and summaries.

It's a system that awards consistency over long periods yet still provides an incentive to chase something/do just a little bit more. Whilst rankings tables come and go, awards are a permanent achievement and will also be ranked on the leaderboard at a later date.

---
User Prefs
---

Adjusted into a tab system to accommodate for more preferences in the very near future.

A summary of how each cardbox schedule works is now viewable in this panel.

---
Chat
---

User's name now shows above any comments made. On occasion there are multiple users on with no FB photo and this was done to remove any confusion this may cause.

Server time added to bottom. This was to reduce confusion where rankings are involved and when the start and end of the day actually occur on Xerafin.

Structurally the chat content is quite different. You'll get used to it.

---
Milestones
---

There's a noticeable chime every time you hit a milestone in chat, There's also a small colour animation at the top of cardbox info, if that panel is active.

We wish to include more audio events in the future and the option of turning them on or off/adjusting the volume. We think this is a good start.

We'd like to thank those who were invited early to test these changes before they went in. Particularly those who reined me in on the color palette/styling where I went a bit too far, we're both very grateful for your feedback.

Happy Xerfing!

Chris & Rick